Better Cluster Regen
Previously clusters only regenerated when we hit a cluster node. Now we properly regenerate whenever we hit a cluster entrance.
There's also a new buffer for when a walkable node changes on an entrance. We still need to regenerate those entrances in case we find a new better entrance position (or the current one became worse).